+++ topobank/analysis/apps.py
@@ -5,5 +5,8 @@ class AnalysisAppConfig(AppConfig):
     name = 'topobank.analysis'
 
     def ready(self):
-        # make sure the signals are registered now
+        # Make sure the signals are registered now
+        # Make sure Celery tasks are registered now
+        from . import custodian  # noqa: F401
         from . import signals  # noqa: F401
+        from . import tasks  # noqa: F401