+++ topobank/settings/base.py
@@ -49,7 +49,7 @@ USE_TZ = True
 # https://docs.djangoproject.com/en/dev/ref/settings/#databases
 postgres_db = env("POSTGRES_DB", default=None)
 if postgres_db is None:
-    DATABASES = {"default": env.db("DATABASE_URL", default="postgres:///topobank-test")}
+    DATABASES = {"default": env.db("DATABASE_URL", default="postgres://postgres@localhost/topobank-test")}
 else:
     DATABASES = {
         "default": {
@@ -330,18 +330,6 @@ CELERY_WORKER_CANCEL_LONG_RUNNING_TASKS_ON_CONNECTION_LOSS = True
 # https://docs.celeryproject.org/en/stable/userguide/configuration.html?highlight=heartbeat#broker-heartbeat
 CELERY_BROKER_HEARTBEAT = 60
 CELERY_REDIS_BACKEND_HEALTH_CHECK_INTERVAL = 30
-
-# https://docs.celeryq.dev/en/latest/userguide/periodic-tasks.html
-CELERY_BEAT_SCHEDULE = {
-    "manager": {
-        "task": "topobank.manager.custodian.periodic_cleanup",
-        "schedule": 12 * 3600,  # Twice a day
-    },
-    "analysis": {
-        "task": "topobank.analysis.custodian.periodic_cleanup",
-        "schedule": 12 * 3600,  # Twice a day
-    },
-}
 
 # django-allauth
 # ------------------------------------------------------------------------------